feat: add report-submit-button
